requirements.txt
alembic==1.9.2
asttokens==2.0.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
executing==0.8.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.3
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
pure-eval==0.2.2
PyMySQL==1.0.2
pysnowflake==0.1.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.0
stack-data==0.3.0
tornado==6.2
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.12.0
